changed project structure to src-layout
