+++ tox.ini
@@ -4,9 +4,9 @@ skip_missing_interpreters = true
 
 [testenv]
 deps = -rrequirements.dev.txt
-changedir = tests
+changedir = test
 commands =
-	pytest --cov=cryo --cov-report=html .
+	pytest --cov=cryostasis --cov-report=html --cov-report=term .
 
 [testenv:pre-commit]
 skip_install = true
